+++ sqlserver/tests/compose-windows/setup.sql
@@ -32,6 +32,12 @@ EXEC sp_addrolemember 'db_datareader', 'fred'
 EXEC sp_addrolemember 'db_datawriter', 'bob'
 GO
 
+-- create an offline database to have an unavailable database to test with
+CREATE DATABASE unavailable_db;
+GO
+ALTER DATABASE unavailable_db SET OFFLINE;
+GO
+
 -- create test procedure for metrics loading feature
 USE master;
 GO
@@ -53,3 +59,17 @@ BEGIN
 END;
 GO
 GRANT EXECUTE on pyStoredProc to datadog;
+GO
+
+CREATE PROCEDURE exampleProcWithoutNocount AS
+BEGIN
+    CREATE TABLE #Hello
+    (
+        [value] int not null,
+    )
+    INSERT INTO #Hello VALUES (1)
+    select * from #Hello;
+END;
+GO
+GRANT EXECUTE on exampleProcWithoutNocount to datadog;
+GO